refactor(core-weave): extract knop inventory renderers
